backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.34.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==2.1.2
google-cloud-storage==2.16.0
python-dotenv==1.0.1
requests==2.32.3
google-cloud-aiplatform==1.56.0